add support for MSP 2; other adjustments
Based on recent interactions it looks like the need for a tool like
this one will remain indefinitely. Oh well.

The refit tool is updated to look for the MSP2 blocks if they are
present, and otherwise load the MSP blocks if those are present
instead. It will not save you from having both packs installed, that
remains your responsibility to prevent.

We also add timestamps to all output so you can see how long each step
takes for your computer.